scripts improvement

+++ deploy/run.sh
@@ -2,10 +2,6 @@
 
 #!/bin/bash
 
-if [ "$#" -ne 1 ]; then
-    echo " Illegal number of parameters, usage: ./run.sh <job name>"
-    exit 1
-fi
 
 XGTOPW_PATH="/home/mateusz/workspace/gtopw/build/Release/"
 PHOTO_TD_PATH="/home/mateusz/workspace/photo_time_dep/build/Release/"
@@ -13,14 +9,25 @@ PHOTO_TD_PATH="/home/mateusz/workspace/photo_time_dep/build/Release/"
 #  change this if you wish [1 - spherical, 0 - cartesian]
 REPRESENTATION=1
 
+#  change this if you wish [1 - delete, 0 - no delete]
+DELETE_INTS=0
+
+
+
+if [ "$#" -ne 1 ]; then
+    echo " Illegal number of parameters, usage: ./run.sh <job name>"
+    exit 1
+fi
 
 HERE=$(dirname $(readlink -f $0))
-mkdir ints inps logs
 INTS="$HERE/ints"
 INPS="$HERE/inps"
+PLOTS="$HERE/plots"
 LOGS="$HERE/logs"
 SETTINGS_FILE="settings.inp"
 JOB_NAME=$1
+
+mkdir $INPS $LOGS
 
 sed -i -e "s/^JOB_NAME .*/JOB_NAME                        ${JOB_NAME}/g" $SETTINGS_FILE
 sed -i -e "s|^RESOURCES_PATH .*|RESOURCES_PATH                        ${INTS}|g" $SETTINGS_FILE
@@ -41,22 +48,32 @@ fi
 
 mv $SETTINGS_FILE $INPS
 
+
+if [ ! -d "$INTS" ]; then
+    echo " Computing ints."
+
+    mkdir $INTS
+    cd $INPS
+    $PHOTO_TD_PATH./main $SETTINGS_FILE -prep >$LOGS/log_prep.out
+
+    if [ "$?" -ne 0 ]; then
+        echo " Preparation of xgtopw input failed."
+        exit 1
+    fi
+
+    sed -i "1i \$PATH\n$INTS/\n\$END" $JOB_NAME.inp
+    $XGTOPW_PATH./xgtopw $JOB_NAME.inp >$LOGS/log_xgtopw.out
+
+    if [ "$?" -ne 0 ]; then
+        echo " Integrals computation (xgtopw) failed."
+        exit 1
+    fi
+
+fi
+
 cd $INPS
-$PHOTO_TD_PATH./main $SETTINGS_FILE -prep >$LOGS/log_prep.out
 
-if [ "$?" -ne 0 ]; then
-    echo " Preparation of xgtopw input failed."
-    exit 1
-fi
-
-sed -i "1i \$PATH\n$INTS/\n\$END" $JOB_NAME.inp
-$XGTOPW_PATH./xgtopw $JOB_NAME.inp >$LOGS/log_xgtopw.out
-
-if [ "$?" -ne 0 ]; then
-    echo " Integrals computation (xgtopw) failed."
-    exit 1
-fi
-
+echo " Computing propagation."
 
 #velocityA gauge
 sed -i -e "s|^GAUGE .*|GAUGE                        velocity_with_Asqrt|g" $SETTINGS_FILE
@@ -95,6 +112,13 @@ fi
 
 
 cd $HERE
-rm -r $INTS
+
+if [ "$DELETE_INTS" -eq 1 ]; then
+    echo " Deleting integrals."
+    rm -r $INTS
+fi
 
 ./plot.py $JOB_NAME
+
+mkdir $PLOTS
+mv *.png $PLOTS
